requirements.txt
# Core dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
chromadb>=0.5.0
tiktoken>=0.7.0
python-dotenv>=1.0.1

# Streamlit app
streamlit>=1.38.0

# Optional (for GPU / local embeddings)
sentence-transformers>=3.0.1
torch>=2.3.0

# PDF + text loaders
pypdf>=4.0.0

# Utility
requests>=2.32.0
